requirements.txt
aiofiles==0.4.0
aiohttp==3.6.2
aiotask-context==0.6.1
argparse==1.4.0
asyncio==3.4.3
asynctest==0.13.0
coverage==5.0.3
flake8==3.7.9
jsonschema==3.2.0
logfmt==0.4
mypy==0.761
PyYAML==5.3
requests==2.22.0
tornado==6.0.3
aiosqlite==0.17.0
asyncio==3.4.3